+++ packages/tldraw/src/state/TLDrawState.ts
@@ -15,7 +15,6 @@ import {
   TLWheelEventHandler,
   Utils,
   TLBounds,
-  Inputs,
 } from '@tldraw/core'
 import {
   FlipType,
@@ -133,6 +132,8 @@ export class TLDrawState extends StateManager<TLDrawSnapshot> {
   private tools = createTools(this)
 
   currentTool: BaseTool = this.tools.select
+
+  editingStartTime = -1
 
   private isCreating = false
 
@@ -486,6 +487,7 @@ export class TLDrawState extends StateManager<TLDrawSnapshot> {
    * @param id [string]
    */
   setEditingId = (id?: string) => {
+    this.editingStartTime = Date.now()
     this.patchState(
       {
         document: {
@@ -2527,14 +2529,15 @@ export class TLDrawState extends StateManager<TLDrawSnapshot> {
   }
 
   onShapeBlur = () => {
+    // This prevents an auto-blur event from Safari
+    if (Date.now() - this.editingStartTime < 50) return
+
     const { editingId } = this.pageState
 
     if (editingId) {
       // If we're editing text, then delete the text if it's empty
       const shape = this.getShape(editingId)
-
       this.setEditingId()
-
       if (shape.type === TLDrawShapeType.Text) {
         if (shape.text.trim().length <= 0) {
           this.setState(Commands.deleteShapes(this.state, [editingId]), 'delete_empty_text')