+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.10
+
+- Fixes spacing in text and sticky shapes.
+
+## 0.1.7
+
+- Fixes text and sticky shapes on iOS.
+
 ## 0.1.4
 
 - UI bug fixes.